updated structure of project so the top NavBar is now in the correct component and it's been imported into the new landing page component
