--- src/components/HeroSection.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import '../App.css';
-import './HeroSection.css';
-
-const HeroSection = () => {
-  return (
-    <header className='landing-container'>
-      <h1>first landing page</h1>
-      <p className='subtext'>look</p>
-    </header>
-  );
-}
-
-export default HeroSection;
