+++ src/components/Landing_Page/Hero_Section/HeroSection.jsx
@@ -1,0 +1,14 @@
+
+
+const HeroSection = () => {
+
+    return (
+        <div>
+            <div>
+                <p>This is the landing page hero section</p>
+            </div>
+        </div>
+    );
+};
+
+export default HeroSection;
