+++ src/context/MyProvider.js
@@ -1,5 +1,5 @@
-import React, { useState } from "react";
-import { useNavigate } from 'react-router-dom';
+import React from "react";
+// import { useNavigate } from 'react-router-dom';
 
 export const MyContext = React.createContext();
